Add template for invitation email
