+++ stories/views.py
@@ -2,43 +2,14 @@
 from __future__ import unicode_literals
 
 # Create your views here.
-from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.core.serializers import serialize
 from django.http import HttpResponse
-from django.utils.decorators import method_decorator
-from django.views.generic import CreateView, TemplateView, FormView
-from django.utils.translation import ugettext_lazy as _
+from django.views.generic import CreateView, TemplateView
 from braces.views import AjaxResponseMixin
 
-from stories.forms import StoryForm, InviteForm
+from stories.forms import StoryForm
 from stories.models import Story
-
-class SendInvite(FormView):
-    template_name = 'invitations/forms/_invite.html'
-    form_class = InviteForm
-
-    @method_decorator(login_required)
-    def dispatch(self, request, *args, **kwargs):
-        return super(SendInvite, self).dispatch(request, *args, **kwargs)
-
-    def form_valid(self, form):
-        email = form.cleaned_data["email"]
-
-        try:
-            invite = form.save(email)
-            invite.inviter = self.request.user
-            invite.save()
-            invite.send_invitation(self.request)
-        except Exception:
-            return self.form_invalid(form)
-        return self.render_to_response(
-            self.get_context_data(
-                success_message=_('%(email)s has been invited') % {
-                    "email": email}))
-
-    def form_invalid(self, form):
-        return self.render_to_response(self.get_context_data(form=form))
 
 
 class StoryCreateView(LoginRequiredMixin, CreateView):
